% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garmintrackR.R
\name{processGarminRunning}
\alias{processGarminRunning}
\title{processGarminRunning}
\usage{
processGarminRunning(data)
}
\arguments{
\item{data}{downloaded from Garmin Connect in .csv format and read into R using read.csv or read.table for example.}
}
\value{
dataframe
}
\description{
A series of functions to process, analyse and plot Garmin Connect data.
}
\examples{
#let's process my data so that it's good for R, and tidy for plotting
my_runs <- processGarminRunning(data=garmin)
summary(my_runs)

}
